+++ discord/commands/holoDeck.js
@@ -1,5 +1,5 @@
 const send = require(`../actions/send`)
-const { log } = require(`../botcommon`)
+const { log, canEdit } = require(`../botcommon`)
 const { usageTag } = require(`../../common`)
 const awaitReaction = require(`../actions/awaitReaction`)
 const Discord = require(`discord.js-light`)
@@ -8,6 +8,7 @@ const depart = require(`../actions/depart`)
 
 module.exports = {
   tag: `holoDeck`,
+  pm: true,
   documentation: false,
   test(content, settings) {
     return new RegExp(`^${settings.prefix}(?:holodeck|holo)$`, `gi`).exec(
@@ -15,7 +16,7 @@ module.exports = {
     )
   },
   async action({ msg, guild }) {
-    log(msg, `Holo Deck`, msg.guild.name)
+    log(msg, `Holo Deck`, msg.guild?.name)
 
     const embed = new Discord.MessageEmbed()
       .setColor(APP_COLOR)
@@ -99,6 +100,6 @@ module.exports = {
       commandsLabel: `Holo Commands`,
       respondeeFilter: (user) => user.id === msg.author.id,
     })
-    sentMessage.delete()
+    if (await canEdit(sentMessage)) sentMessage.delete()
   },
 }
